Add a possibility to set the value of the search field
This can be useful if you want to pre-fill a Search Form.

For example, when adding a file to a database, you may want to ask the user to first provide the file, then some JS logic may enable you to guess the category the file belongs to. The search field about the category would then be pre-filled, enabling the user to select faster the category he meant.